+++ src/main/java/cn/edu/thssdb/query/DropTableStatement.java
@@ -1,6 +1,7 @@
 
 package cn.edu.thssdb.query;
 
+import cn.edu.thssdb.exception.CurrentDatabaseNullException;
 import cn.edu.thssdb.parser.SQLParser.Drop_table_stmtContext;
 import cn.edu.thssdb.parser.SQLParser.Sql_stmtContext;
 import cn.edu.thssdb.rpc.thrift.ExecuteStatementResp;
@@ -25,6 +26,9 @@ public class DropTableStatement extends Statement {
 
     @Override
     public final void execute() {
+        if (manager.currentDatabase == null) {
+            throw new CurrentDatabaseNullException();
+        }
         Database db = this.manager.currentDatabase;
         db.dropTable(this.tablename);
         result = constructSuccessResp(String.format("drop table %s success!", this.tablename));
